agent-ppt/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pppt"
version = "0.2.0"
description = "JSON to PPTX exporter via Open XML"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=9.0",
]

[project.optional-dependencies]
svg = ["cairosvg>=2.5"]
dev = ["pytest>=7.0"]

[project.scripts]
pppt = "pppt.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
